requirements.txt
watchdog
IMDbPY
strsim